+++ resources/inertia/Pages/Validators.contracts.ts
@@ -1,4 +1,27 @@
-export interface IStatistics {
+import { IValidator } from '../types';
+
+export interface IValidatorData {
+    statistics: IMonitorStatistics;
+    overflowValidators: IValidator[];
+    validators: IValidator[];
+}
+
+export interface IMonitorStatistics {
+    performances?: {
+        forging?: string | number;
+        missed?: string | number;
+        missing?: string | number;
+    };
+    blockCount?: number;
+    nextValidator?: {
+        address?: string;
+        attributes?: {
+            username?: string;
+        };
+    };
+}
+
+export interface IValidatorsStatistics {
     voterCount: number;
     totalVoted: number;
     missedBlocks: number;
@@ -7,5 +30,5 @@ export interface IStatistics {
 }
 
 export interface ValidatorsProps {
-    statistics: IStatistics;
+    statistics: IValidatorsStatistics;
 }
