+++ resources/inertia/Pages/Validators/Monitor.tsx
@@ -6,12 +6,12 @@ import HeaderStats from "@/Components/Validator/Monitor/HeaderStats";
 import ValidatorFavoritesProvider from "@/Providers/ValidatorFavorites/ValidatorFavoritesProvider";
 import MonitorMobileTableWrapper from "@/Components/Tables/Mobile/Validators/Monitor";
 import MobileDivider from "@/Components/General/MobileDivider";
-import { IValidatorData } from "@/types";
 import MissedBlocksTrackerProvider from "@/Providers/MissedBlocksTracker/MissedBlocksTrackerProvider";
 import { useTranslation } from "react-i18next";
 import { usePageMetadata } from "@/Components/General/Metadata";
 import { PageProps } from "@inertiajs/core";
 import Layout from "@/Layout";
+import { IValidatorData } from "../Validators.contracts";
 
 export default function Monitor({
     validatorData,
